new change

+++ server/controllers/controller_cause_list.js
@@ -78,14 +78,38 @@ module.exports = {
     },
     updateCauseDataById: async (req, res)=> {
         try {
-            const data = await CauseSchema.findById(req.params.id)
-            data.title = req.body.title
-            data.description = req.body.description
-            res.status(201).json({
-                success: true,
-                message: 'Data Updated successfully',
-                Cause: data,
-            });
+            if(req.headers.authorization && req.headers.authorization.startsWith('Bearer ')){
+                token = req.headers.authorization.split(" ")[1] 
+                let {userId, date, iat} = commonFunctions.decryptJwt(token)
+                const {userId: idd} = await CauseSchema.findById(req.params.id)
+                console.log( ">>", userId,  idd.toString())
+
+                if (idd.toString() == userId) {
+                    const data = await CauseSchema.findById(req.params.id)
+                    
+                    data.title = req.body.title
+                    data.description = req.body.description
+
+                    const createdCause = await data.save()
+                    res.status(200).json({
+                        success: true,
+                        message: 'Data Updated successfully',
+                        createdCause
+                    })
+
+                } else {                    
+                    res.status(400).json({
+                        success: true,
+                        message: 'Unauthorized access',
+                    })
+                }
+            } else {
+                res.status(400).json({
+                    success: true,
+                    message: 'Unauthorized access',
+                })
+            }
+            
         } catch(err) {
             res.status(500).json({
                 success: false,
@@ -97,12 +121,30 @@ module.exports = {
     
     deleteCauseDataById: async (req, res)=> {
         try {
-            const data = await CauseSchema.findByIdAndRemove(req.params.id)
-            res.status(201).json({
-                success: true,
-                message: 'Data Deleted Successfully',
-                Cause: data,
-            });
+            if(req.headers.authorization && req.headers.authorization.startsWith('Bearer ')){
+                token = req.headers.authorization.split(" ")[1] 
+                let {userId, date, iat} = commonFunctions.decryptJwt(token)
+                const {userId: idd} = await CauseSchema.findById(req.params.id)
+                console.log( ">>", userId,  idd)
+
+                if (idd.toString() == userId) {
+                    const data = await CauseSchema.findByIdAndRemove(req.params.id)
+                    res.status(201).json({
+                        success: true,
+                        message: 'Data Deleted Successfully',
+                        Cause: data,
+                    });
+                    console.log("condition success")
+                } 
+                else {
+                    console.log("condition fail")
+                    res.status(400).json({
+                        success: true,
+                        message: 'Unauthorized access',
+                    })
+                }
+            }
+
         } catch(err) {
             res.status(500).json({
                 success: false,
